[QuantizationModifier] refactor QuantizationScheme to its own file (#1223)
