+++ src/sparseml/pytorch/sparsification/quantization/helpers.py
@@ -18,7 +18,6 @@ Helper functions for performing quantization aware training with PyTorch
 
 from copy import deepcopy
 from dataclasses import dataclass, field
-from functools import partial
 from typing import Any, Callable, Dict, List, Optional, Tuple, Union
 
 import torch
@@ -28,11 +27,13 @@ from torch import quantization as torch_quantization
 from torch.nn import BatchNorm2d, Conv2d, Embedding, Module, ReLU
 
 from sparseml.pytorch.nn import ReLU as ReLU_nm
+from sparseml.pytorch.sparsification.quantization.quantization_scheme import (
+    get_observer,
+)
 from sparseml.pytorch.utils import get_layer
 
 
 _PARSED_TORCH_VERSION = version.parse(torch.__version__)
-_TORCH_PRE_112 = _PARSED_TORCH_VERSION < version.parse("1.12.0")
 
 __all__ = [
     "QATWrapper",
@@ -444,26 +445,6 @@ def configure_module_qat_wrappers(
         )
 
 
-def compute_range(dtype: torch.dtype, bits: int):
-    """
-    compute quantization limits depending on data type and number of bits
-
-    :param dtype: data type.
-    :param bits: number of bits.
-    :return: minimum limit, maximum limit, whether the range is customized
-    """
-    bits = bits if bits else 8
-    is_custom = bits != 8
-    if dtype == torch.qint8:
-        quant_min = -(2 ** (bits - 1))
-        quant_max = (2 ** (bits - 1)) - 1
-    elif dtype == torch.quint8:
-        quant_min = 0
-        quant_max = (2 ** bits) - 1
-
-    return quant_min, quant_max, is_custom
-
-
 def configure_module_default_qconfigs(module: Module):
     """
     if any submodule of the given module has a configure_qconfig function,
@@ -562,75 +543,6 @@ def get_qat_qconfig(qproperties: QConfigProperties) -> "torch.quantization.QConf
         activation=activation_observer,
         weight=weight_observer,
     )
-
-
-def get_observer(
-    symmetric: bool,
-    dtype: torch.dtype,
-    bits: int,
-    reduce_range: bool,
-    qconfig_kwargs: Dict[str, Any],
-):
-    qscheme = torch.per_tensor_symmetric if symmetric else torch.per_tensor_affine
-    quant_min, quant_max, is_custom_qrange = compute_range(dtype, bits)
-
-    observer_cls = torch_quantization.MovingAverageMinMaxObserver
-    observer_kwargs = dict(
-        dtype=dtype,
-        qscheme=qscheme,
-        reduce_range=reduce_range,
-    )
-
-    """
-    in torch 1.9.1, quant_min and quant_max are not passed to observer:
-    https://github.com/pytorch/pytorch/blob/v1.9.1/torch/quantization/fake_quantize.py#L109
-    however in 1.12.0, this is fixed so both are passed to observer:
-    https://github.com/pytorch/pytorch/blob/v1.12.1/torch/ao/quantization/fake_quantize.py#L132
-
-    Passing quant_min/quant_max to observer means the observer will have
-    `self.has_customized_qrange == True` in both 1.9.1 and 1.12.0.
-
-    For whatever reason, both versions calculate zero point for
-    quint8 differently **if there is a customized_qrange**
-    1. customized qrange has zero point of 127
-    2. non-customized has zero point of 128.
-    source:
-    https://github.com/pytorch/pytorch/blob/v1.12.1/torch/ao/quantization/observer.py#L293
-
-    **we want to ensure that the zero point is 128**
-    see https://github.com/neuralmagic/sparseml/pull/604
-    """
-    if is_custom_qrange:
-        # for both versions we need to include the custom min/max values in kwargs
-        observer_kwargs["quant_min"] = quant_min
-        observer_kwargs["quant_max"] = quant_max
-        if _TORCH_PRE_112:
-            # pre 1.12, the observer doesn't get passed the quant_min/quant_max values,
-            # so we patch them in to the constructor of the observer
-            observer_cls = partial(
-                observer_cls, quant_min=quant_min, quant_max=quant_max
-            )
-    else:
-        # if using a non custom qrange, we can rely on default values used by
-        # the observers
-        if _TORCH_PRE_112:
-            # pre 1.12, the observer doesn't get passed the quant_min/quant_max values,
-            # so we are safe to pass these to FakeQuantize
-            observer_kwargs["quant_min"] = quant_min
-            observer_kwargs["quant_max"] = quant_max
-        else:
-            # post 1.12 we cannot pass them to the observer since that will set
-            # has_customized_qrange. instead we rely on the default values
-            # being equal to the `quant_min` and `quant_max` here.
-            pass
-
-    observer_kwargs["observer"] = observer_cls
-    observer_kwargs.update(qconfig_kwargs or {})
-    observer = torch_quantization.FakeQuantize.with_args(
-        **observer_kwargs,
-    )
-
-    return observer
 
 
 def freeze_bn_stats(module: Module):
